requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
tensorflow==2.14.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
torch==2.1.1
torchvision==0.16.1
yolov5==7.0.13
huggingface-hub==0.19.4
